Fix telemetry lint failures
Move test-only fixtures out of production code and keep Linux mount parsing behind its platform build constraint.
